+++ src/main/java/com/flowiee/pms/repository/sales/MaterialTempRepository.java
@@ -12,20 +12,20 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 @Repository
-public interface MaterialTempRepository extends JpaRepository<MaterialTemp, Integer> {
+public interface MaterialTempRepository extends JpaRepository<MaterialTemp, Long> {
     @Query("from MaterialTemp m where m.ticketImport.id=:importId")
-    List<MaterialTemp> findByImportId(@Param("importId") Integer importId);
+    List<MaterialTemp> findByImportId(@Param("importId") Long importId);
     
     @Query("from MaterialTemp m where m.ticketImport.id=:importId and m.material.id=:materialId")
-    MaterialTemp findMaterialInGoodsImport(@Param("importId") Integer importId, @Param("materialId") Integer materialId);
+    MaterialTemp findMaterialInGoodsImport(@Param("importId") Long importId, @Param("materialId") Long materialId);
 
     @Transactional
     @Modifying
     @Query("update MaterialTemp m set m.quantity = (m.quantity + :quantity) where m.id =:materialTempId")
-    void updateQuantityIncrease(@Param("materialTempId") Integer materialTempId, @Param("quantity") Integer quantity);
+    void updateQuantityIncrease(@Param("materialTempId") Long materialTempId, @Param("quantity") Integer quantity);
 
     @Transactional
     @Modifying
     @Query("update MaterialTemp m set m.storageQty=:storageQty where m.id =:materialTempId")
-    void updateStorageQuantity(@Param("materialTempId") Integer materialTempId, @Param("storageQty") Integer storageQty);
+    void updateStorageQuantity(@Param("materialTempId") Long materialTempId, @Param("storageQty") Integer storageQty);
 }
