+++ src/main/java/com/flowiee/pms/repository/sales/TicketImportRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TicketImportRepository extends JpaRepository<TicketImport, Integer> {
+public interface TicketImportRepository extends JpaRepository<TicketImport, Long> {
     @Query("from TicketImport i " +
            "where (:text is null or i.title like %:text%) " +
            "and (:supplierId is null or i.supplier.id=:supplierId) " +
@@ -18,13 +18,13 @@ public interface TicketImportRepository extends JpaRepository<TicketImport, Inte
            "and (:importStatus is null or i.status=:importStatus) " +
            "and (:storageId is null or i.storage.id=:storageId)")
     Page<TicketImport> findAll(@Param("text") String text,
-                               @Param("supplierId") Integer supplierId,
-                               @Param("paymentMethod") Integer paymentMethod,
+                               @Param("supplierId") Long supplierId,
+                               @Param("paymentMethod") Long paymentMethod,
                                @Param("payStatus") String payStatus,
                                @Param("importStatus") String importStatus,
-                               @Param("storageId") Integer storageId,
+                               @Param("storageId") Long storageId,
                                Pageable pageable);
 
     @Query("from TicketImport i where i.status=:status and i.createdBy=:createdBy")
-    TicketImport findDraftGoodsImportPresent(@Param("status") String status, @Param("createdBy") Integer createdBy);
+    TicketImport findDraftGoodsImportPresent(@Param("status") String status, @Param("createdBy") Long createdBy);
 }
